refactor: 검증 로직 객체 분리
